+++ DKDBManager/NSManagedObject+DKDBManager.m
@@ -23,17 +23,36 @@ void        CRUDLog(BOOL logEnabled, NSString * _Nonnull format, ...) {
 
 @implementation NSManagedObject (DKDBManager)
 
-#pragma mark - CREATE
+#pragma mark - Deprecated functions
 
 + (instancetype _Nullable)createEntityInContext:(NSManagedObjectContext * _Nonnull)savingContext {
-	return [self createEntityFromDictionary:NSDictionary.new inContext:savingContext completion:nil];
+	return [self crudEntityInContext:savingContext];
 }
 
 + (instancetype _Nullable)createEntityFromDictionary:(NSDictionary * _Nullable)dictionary inContext:(NSManagedObjectContext * _Nonnull)savingContext {
-    return [self createEntityFromDictionary:dictionary inContext:savingContext completion:nil];
+	return [self crudEntityWithDictionary:dictionary inContext:savingContext];
 }
 
 + (instancetype _Nullable)createEntityFromDictionary:(NSDictionary * _Nullable)dictionary inContext:(NSManagedObjectContext * _Nonnull)savingContext completion:(void (^ _Nullable)(id _Nullable entity, DKDBManagedObjectState status))completion {
+	return [self crudEntityWithDictionary:dictionary inContext:savingContext completion:completion];
+}
+
++ (NSArray * _Nullable)createEntitiesFromArray:(NSArray * _Nonnull)array inContext:(NSManagedObjectContext * _Nonnull)savingContext {
+	return [self crudEntitiesWithArray:array inContext:savingContext];
+}
+
+#pragma mark - CREATE
+
+
++ (instancetype _Nullable)crudEntityInContext:(NSManagedObjectContext * _Nonnull)savingContext {
+	return [self crudEntityWithDictionary:NSDictionary.new inContext:savingContext completion:nil];
+}
+
++ (instancetype _Nullable)crudEntityWithDictionary:(NSDictionary * _Nullable)dictionary inContext:(NSManagedObjectContext * _Nonnull)savingContext {
+    return [self crudEntityWithDictionary:dictionary inContext:savingContext completion:nil];
+}
+
++ (instancetype _Nullable)crudEntityWithDictionary:(NSDictionary * _Nullable)dictionary inContext:(NSManagedObjectContext * _Nonnull)savingContext completion:(void (^ _Nullable)(id _Nullable entity, DKDBManagedObjectState status))completion {
 
     // now create, save or update an entity
     DKDBManagedObjectState status = DKDBManagedObjectStateSave;
@@ -64,15 +83,14 @@ void        CRUDLog(BOOL logEnabled, NSString * _Nonnull format, ...) {
     return entity;
 }
 
-+ (NSArray * _Nullable)createEntitiesFromArray:(NSArray * _Nonnull)array inContext:(NSManagedObjectContext * _Nonnull)savingContext {
-
++ (NSArray * _Nullable)crudEntitiesWithArray:(NSArray * _Nonnull)array inContext:(NSManagedObjectContext * _Nonnull)savingContext {
 	if (array == nil || array.count == 0) {
 		return nil;
 	}
 
     NSMutableArray *entities = [NSMutableArray new];
     for (NSDictionary *dictionary in array) {
-        id entity = [self createEntityFromDictionary:dictionary inContext:savingContext];
+        id entity = [self crudEntityWithDictionary:dictionary inContext:savingContext];
         if (entity != nil) {
             [entities addObject:entity];
         }